requirements.txt
# installing required libraries in my_env
langchain==0.1.11 
gradio==5.23.2 
transformers==4.38.2 
bs4==0.0.2 
requests==2.31.0 
torch
virtualenv==20.24.0
pillow==10.4.0